Archiver And SendRoute Rule Export CURL sh
This uses curl bash on linux to auto export and date send Route rules
from sterling integrator LightWell Framework in XML that can later be
imported.  Archiver moves to an ftp site.

Working on automated importer to be able to keep a history of rules
files and schedule imports into dev to sync prod and dev systems